refactor: 최근 1시간 이내 미배정 작업자 조회 로직 변경

+++ src/main/java/com/detailretail/kurlyflow/admin/util/CalculateConverter.java
@@ -2,7 +2,7 @@ package com.detailretail.kurlyflow.admin.util;
 
 public class CalculateConverter {
 
-  public static long getSeventy(Integer workingNumbers) {
-    return Math.round(((double) workingNumbers / 100) * 70);
+  public static int getSeventy(Integer workingNumbers) {
+    return (int) Math.round(((double) workingNumbers / 100) * 70);
   }
 }
